cambia: extrae la funcion checkIfUserHasTheDefaultBoard en un archivo propio
